Move contextual folder to root
