+++ src/perf.js
